Minimal - responsive Landing page
Landing page is built using HTML, CSS, JavaScript.
This landing is responsive to different devices, such as tablets and smartphones, where the menu automatically collapses into a burger menu. In addition, when updating the page, animations are used, which makes the process of the appearance of text and images more attractive to the user.